requirement.txt
aniso8601==9.0.1
click==7.1.2
cycler==0.10.0
dbfread==2.0.7
dnspython==1.16.0
eventlet==0.33.3
Flask==1.1.4
Flask-RESTful==0.3.10
Flask-SocketIO==2.8.4
functools32==3.2.3.post2
future==0.15.2
greenlet==2.0.2
inotify==0.2.10
itsdangerous==1.1.0
Jinja2==2.11.3
logging==0.4.9.6
MarkupSafe==1.1.1
matplotlib==1.5.2
monotonic==1.6
mpmath==0.19
nose==1.3.7
numpy==1.9.3
pandas==0.18.1
prophet==0.1.1.post1
pyparsing==2.1.4
pyshp==2.3.1
python-dateutil==2.5.3
python-engineio==3.2.0
python-socketio==4.0.1
pytz==2016.6.1
PyYAML==5.3.1
requests==2.11.1
scipy==0.17.0
six==1.10.0
sympy==1.0
Werkzeug==1.0.1
xlrd==1.0.0
xlwt==1.1.2
